Updated teensy to use common code from stm directory.
Updated mconfigport.h to enable GC
