+++ stm/Makefile
@@ -30,7 +30,8 @@ SRC_C = \
 	string0.c \
 	malloc0.c \
 	systick.c  \
-	lexerstm.c \
+	lexerstr.c \
+	lexerfatfs.c \
 	led.c \
 	lcd.c \
 	servo.c \
