+++ stm/lexerstr.h
