+++ stm/main.c
@@ -19,7 +19,8 @@
 #include "nlr.h"
 #include "misc.h"
 #include "lexer.h"
-#include "lexerstm.h"
+#include "lexerstr.h"
+#include "lexerfatfs.h"
 #include "parse.h"
 #include "obj.h"
 #include "compile.h"
