+++ teensy/led.c
@@ -61,21 +61,17 @@ mp_obj_t led_obj_off(mp_obj_t self_in) {
 static MP_DEFINE_CONST_FUN_OBJ_1(led_obj_on_obj, led_obj_on);
 static MP_DEFINE_CONST_FUN_OBJ_1(led_obj_off_obj, led_obj_off);
 
+static const mp_method_t led_methods[] = {
+    { "on", &led_obj_on_obj },
+    { "off", &led_obj_off_obj },
+    { NULL, NULL },
+};
+
 static const mp_obj_type_t led_obj_type = {
     { &mp_const_type },
     "Led",
-    led_obj_print, // print
-    NULL, // make_new
-    NULL, // call_n
-    NULL, // unary_op
-    NULL, // binary_op
-    NULL, // getiter
-    NULL, // iternext
-    { // method list
-        { "on", &led_obj_on_obj },
-        { "off", &led_obj_off_obj },
-        { NULL, NULL },
-    }
+    .print = led_obj_print,
+    .methods = led_methods,
 };
 
 mp_obj_t pyb_Led(mp_obj_t led_id) {
